refactor: Organize bins more like cargo
