+++ src/bin/add/cli.rs
@@ -1,0 +1,22 @@
+use cargo_edit::CargoResult;
+use clap::Parser;
+
+#[derive(Debug, Parser)]
+#[clap(bin_name = "cargo")]
+pub enum Command {
+    Add(crate::add::AddArgs),
+}
+
+impl Command {
+    pub fn exec(self) -> CargoResult<()> {
+        match self {
+            Self::Add(add) => add.exec(),
+        }
+    }
+}
+
+#[test]
+fn verify_app() {
+    use clap::IntoApp;
+    Command::into_app().debug_assert()
+}
